# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ICRM-CA API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/applications": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "List Applications",
        "operationId": "list_applications_api_v1_applications_get",
        "parameters": [
          {
            "name": "icrm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Icrm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationResponse"
                  },
                  "title": "Response List Applications Api V1 Applications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Create Application",
        "operationId": "create_application_api_v1_applications_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "icrm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Icrm Session"
            }
          },
          {
            "name": "icrm_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Icrm Csrf"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationFields"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/applications/{application_id}/redline": {
      "get": {
        "tags": [
          "redline"
        ],
        "summary": "Get Redline",
        "operationId": "get_redline_api_v1_applications__application_id__redline_get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Application Id"
            }
          },
          {
            "name": "icrm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Icrm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiveRedlineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{application_id}/rule-context": {
      "post": {
        "tags": [
          "redline"
        ],
        "summary": "Confirm Rule Context",
        "operationId": "confirm_rule_context_api_v1_applications__application_id__rule_context_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Application Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "icrm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Icrm Session"
            }
          },
          {
            "name": "icrm_csrf",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Icrm Csrf"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleContextConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/applications/{application_id}/redline-runs/{run_id}": {
      "get": {
        "tags": [
          "redline"
        ],
        "summary": "Get Redline Run",
        "operationId": "get_redline_run_api_v1_applications__application_id__redline_runs__run_id__get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Application Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "icrm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Icrm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__redline_api__RunDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{application_id}/redline-runs/{run_id}/printable": {
      "get": {
        "tags": [
          "redline"
        ],
        "summary": "Printable Redline Run",
        "operationId": "printable_redline_run_api_v1_applications__application_id__redline_runs__run_id__printable_get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Application Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "icrm_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Icrm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationFields": {
        "properties": {
          "primary_borrower": {
            "$ref": "#/components/schemas/PrimaryBorrower"
          },
          "product": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Product"
          },
          "application_date": {
            "type": "string",
            "format": "date",
            "title": "Application Date"
          },
          "proposed_signing_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Signing Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "primary_borrower",
          "product",
          "application_date"
        ],
        "title": "ApplicationFields"
      },
      "ApplicationResponse": {
        "properties": {
          "primary_borrower": {
            "$ref": "#/components/schemas/PrimaryBorrower"
          },
          "product": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Product"
          },
          "application_date": {
            "type": "string",
            "format": "date",
            "title": "Application Date"
          },
          "proposed_signing_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Signing Date"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "lifecycle_state": {
            "type": "string",
            "title": "Lifecycle State"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "primary_borrower",
          "product",
          "application_date",
          "id",
          "owner_id",
          "lifecycle_state",
          "version"
        ],
        "title": "ApplicationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuleContextConfirmRequest": {
        "properties": {
          "context": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Context"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "context"
        ],
        "title": "RuleContextConfirmRequest"
      },
...
    }
  }
}